feat: folder organize script

+++ scripts/section-8/folder_organizar.sh
@@ -2,19 +2,35 @@
 
 read -r -p "Which folder do you want to organize? " folder
 
+if [ ! -d "$folder" ]; then
+	echo "Invalid folder provided"
+	exit 1
+fi
+
+
 while read filename; do 
 	case "$filename" in 
-		*.pdf) subfolder="pdf" ;;
-		*.jpg|*.jpge) subfolder="images";;
-		*.mp4) subfolder="videos" ;;
-		*) subfolder="others" ;;
+		*.txt) subfolder="plaintext";;
+		*.pdf) subfolder="pdf";;
+		*.jpg|*.jpeg) subfolder="images";;
+		*.mp4) subfolder="videos";;
+		*) subfolder="others";;
 	esac
+
 	if [ ! -d "$folder/$subfolder" ]; then
 		mkdir -p "$folder/$subfolder"
 	fi
+
 	if [ -d "$filename" ]; then
 		continue
 	fi
-	echo "folder: $folder/$subfolder"
+	
+	if [ -f "$folder/$subfolder/$filename" ]; then
+		continue
+	fi
+
 	mv "$filename" "$folder/$subfolder"
-done < <(ls "$folder")
+done < <(find "$folder" -maxdepth 1 -type f)
+
+
+echo "All files moved!"
